% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_matrix_reduction.r
\name{binary_matrix_reduction}
\alias{binary_matrix_reduction}
\title{Creating a binary existence matrix}
\usage{
binary_matrix_reduction(binary_data, batch_list, needed_values)
}
\arguments{
\item{binary_data}{The input data.frame that should become binary.}

\item{batch_list}{Information about the sample's batch affiliations.}

\item{needed_values}{Information, how many values are needed to render a
a batch 'valid'.}
}
\value{
A binary existence matrix returned as a data.frame
}
\description{
This function reduces its input matrix to a binary existence matrix based on
the given description file (and information on how many values a batch needs)
for proper adjustment.
}
