% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.r
\name{build_key_list}
\alias{build_key_list}
\title{Creation of keys}
\usage{
build_key_list(batch_list)
}
\arguments{
\item{batch_list}{The list with information about batch-affiliations for
every sample.}
}
\value{
A list element with usable keys
}
\description{
Calculates a list of usable keys based on the passed batch listings
}
