% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_row.r
\name{crop_row}
\alias{crop_row}
\title{Cropping current row}
\usage{
crop_row(main_data, batch_data, batch_list, each_vector, current_row)
}
\arguments{
\item{main_data}{This is the input data.frame read in by the HarmonizR.}

\item{batch_data}{This is the description data.frame read in by the
HarmonizR.}

\item{batch_list}{An overview of the batch groupings in list form.}

\item{each_vector}{A single vector extracted from affiliation_list. For
further information see the return value from the spotting_missing_values()
function.}

\item{current_row}{This is the data the function is editing.}
}
\value{
A correctly cropped row for the splitting() function
}
\description{
This function is repeadiately used by the splitting() function to make edits
on the current row from the given sub-data.frame.
}
