% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_description.r
\name{fetch_batch_overview}
\alias{fetch_batch_overview}
\title{Fetching batch list}
\usage{
fetch_batch_overview(batch_data)
}
\arguments{
\item{batch_data}{This is a data.frame and simultaneously the result from
read_description()}
}
\value{
Batch distribution as list
}
\description{
The fetch_batch_overview function extracts the overview over the batch
distribution in list format.
}
