% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.r
\name{find_na}
\alias{find_na}
\title{Finding NAs for the sorting process}
\usage{
find_na(df, batch_list)
}
\arguments{
\item{df}{The data.frame passed initially by the user.}

\item{batch_list}{The list with information about batch-affiliations for
every sample.}
}
\value{
An overview of the NA-distribution
}
\description{
Creates an overview of NAs based on both the passed input data.frame and
the batch list
}
