% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_S4.r
\name{format_from_S4}
\alias{format_from_S4}
\title{Format data taken from S4}
\usage{
format_from_S4(data)
}
\arguments{
\item{data}{Data (S4 format) passed by the user. No description file is
needed when using S4 data}
}
\value{
Data and description as data.frames
}
\description{
This function converts passed S4 summarized experiment data to HarmonizR
input
}
