% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_description.r
\name{read_description}
\alias{read_description}
\alias{fetch_batch_overview}
\title{Reading description}
\usage{
read_description(description_source)

fetch_batch_overview(batch_data)
}
\arguments{
\item{description_source}{Usually the path to the description file. It can
also be a correctly formatted data.frame.}

\item{batch_data}{This is a data.frame and simultaniously the result from
read_description()}
}
\value{
read_description(): Description as data.frame

fetch_batch_overview(): Batch distribution as list
}
\description{
The read_description function reads in a file via its file path and converts
it to a for the rest of the workflow readable format.

The fetch_batch_overview function extracts the overview over the batch
distribution in list format.
}
