% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.r
\name{jaccard}
\alias{jaccard}
\title{Jaccard-based sorting}
\usage{
jaccard(binary_df)
}
\arguments{
\item{binary_df}{The input matrix passed by the user reduced to presence and
absence of features in batches (binary)}
}
\value{
A template for batch-sorting based on Jaccard similarity
}
\description{
Calculates a order to sort by based on the Jaccard similarity of all given
batches
}
