% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.r
\name{jaccard_index_absence}
\alias{jaccard_index_absence}
\title{Jaccard index on zeroes (absence)}
\usage{
jaccard_index_absence(a, b)
}
\arguments{
\item{a}{First list with either 0 or 1 entries to be compared against the
second list.}

\item{b}{Second list with either 0 or 1 entries to be compared against the
first list.}
}
\value{
The Jaccard similarity based on absent values
}
\description{
Calculates the Jaccard index for two given lists a and b based on common
zeroes
}
