% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_description.r
\name{read_description}
\alias{read_description}
\title{Reading description}
\usage{
read_description(description_source)
}
\arguments{
\item{description_source}{Usually the path to the description file. It can
also be a correctly formatted data.frame.}
}
\value{
Description as data.frame
}
\description{
The read_description function reads in a file via its file path and converts
it to a for the rest of the workflow readable format.
}
