% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_main_data.r
\name{read_main_data}
\alias{read_main_data}
\title{Reading main data}
\usage{
read_main_data(data_source)
}
\arguments{
\item{data_source}{Usually the path to the input data. It can also be passed
directly as a correctly formatted data.frame.}
}
\value{
To-be-adjusted data as data.frame
}
\description{
The read_main_data function reads in a file via its file path and converts
it to a for the rest of the workflow readable format.
}
