% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.r
\name{splitting}
\alias{splitting}
\title{Splitting}
\usage{
splitting(
  affiliation_list,
  main_data,
  batch_data,
  batch_list,
  algorithm,
  ComBat_mode
)
}
\arguments{
\item{affiliation_list}{An overview of which protein has which missing value
distribution.}

\item{main_data}{This is the input data.frame read in by the HarmonizR.}

\item{batch_data}{This is the description data.frame read in by the
HarmonizR.}

\item{batch_list}{An overview of the batch groupings in list form.}

\item{algorithm}{Either "ComBat" or "limma". Based on the selected algorithm
for the harmonizR() function.}

\item{ComBat_mode}{The chosen ComBat mode influences the parameters the
ComBat algorithm is using. Based on the ComBat_mode parameter given to the
harmonizR() function. Not active during limma execution.}
}
\value{
Returns a list of 'chopped up' data.frames
}
\description{
This function splits the data.frame. The data is very sensitive to its
specific input. Only to be called via harmonizR()
}
