% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.r
\name{visual}
\alias{visual}
\alias{visual2}
\alias{visual3}
\title{Visualization}
\usage{
visual(input_dataframe, batch_list)

visual2(input_dataframe, batch_list)

visual3(input_dataframe, batch_list)
}
\arguments{
\item{input_dataframe}{A data.frame object as input.}

\item{batch_list}{A list object giving information about which column
corresponds to which batch.}
}
\value{
A data.frame object, which is ready to be plotted
}
\description{
The visual functions turn their input dataframes into easily plottable
results.
}
