% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.r
\name{visual2}
\alias{visual2}
\title{Visualize sample means}
\usage{
visual2(input_dataframe, batch_list)
}
\arguments{
\item{input_dataframe}{A data.frame object as input.}

\item{batch_list}{A list object giving information about which column
corresponds to which batch.}
}
\value{
A data.frame object, which is ready to be plotted
}
\description{
The visual functions turn their input dataframes into easily plottable
results.
}
