% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.r
\name{blocking}
\alias{blocking}
\title{Blocking}
\usage{
blocking(batch_list, block)
}
\arguments{
\item{batch_list}{The list with information about batch-affiliations for
every sample.}

\item{block}{The blocking parameter (how many batches should always get
blocked together).}
}
\value{
Returns an updated 'batch_list' with blocking included
}
\description{
This function performs blocking on the given description and therefore
influences how the dataset will be split later down the pipeline.
}
