% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{harmonizR}
\alias{harmonizR}
\title{Main function}
\usage{
harmonizR(
  data_as_input = NULL,
  description_as_input = NULL,
  ...,
  algorithm = "ComBat",
  ComBat_mode = 1,
  plot = FALSE,
  output_file = "cured_data"
)
}
\arguments{
\item{data_as_input}{Path to input data. Please read the SOP for the correct
input format found on https://github.com/SimonSchlumbohm/HarmonizR.
Additionally, the input can be a data.frame with proper row- and column
names.}

\item{description_as_input}{Path to input description. Please read the SOP
for the correct input format found on
https://github.com/SimonSchlumbohm/HarmonizR. Additionally, the input can be
a data.frame with three columns total.}

\item{algorithm}{Optional. Pass either "ComBat" or "limma" to select the
preferred adjustment method. Defaults to ComBat.}

\item{ComBat_mode}{Optional. Pass a number between 1 and 4 to select the
desired ComBat parameters. Can only be set when ComBat is used. For
information on the meaning of the numbers, please view the SOP found on
https://github.com/SimonSchlumbohm/HarmonizR. Defaults to 1.}

\item{plot}{Optional. Takes either "samplemeans" for sample specific means,
"featuremeans" for feature specific means or "CV" for the coeffivient of
variation as input and creates before/after plots for the given data.
Defaults to FALSE -> Turned off.}

\item{output_file}{Optional. Takes a string as input for the .tsv file name.
This can also be a path. Defaults to "cured_data", hence yielding a
"cured_data.tsv" file.}
}
\value{
A .tsv file by default called cured_data.tsv will be written out as
a result.
As a return, the harmonizR function will yield the batch effect adjusted
data.frame.
}
\description{
This function executes the entire harmonizR program and executes all other
functions found in this package. Therefore, this is the only function in
need of calling.
}
