% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.r
\name{splitting}
\alias{splitting}
\title{Splitting}
\usage{
splitting(
  affiliation_list,
  main_data,
  batch_data,
  block_list,
  algorithm,
  ComBat_mode,
  block,
  verbosity,
  cores
)
}
\arguments{
\item{affiliation_list}{An overview of which protein has which missing value
distribution.}

\item{main_data}{This is the input data.frame read in by the HarmonizR.}

\item{batch_data}{This is the description data.frame read in by the
HarmonizR.}

\item{block_list}{An overview of the batch groupings in list form. If the
block parameter was used, the groupings are changed accordingly.}

\item{algorithm}{Either "ComBat" or "limma". Based on the selected algorithm
for the harmonizR() function.}

\item{ComBat_mode}{The chosen ComBat mode influences the parameters the
ComBat algorithm is using. Based on the ComBat_mode parameter given to the
harmonizR() function. Not active during limma execution.}

\item{block}{The block parameter is here used to determine whether there are
single-batch dataframes at all present.}

\item{verbosity}{Toggles the amount of stuff printed out by the HarmonizR
algorithm during execution.}

\item{cores}{Manually sets the number of cores the user wants to be used
during HarmonizR's execution. A positive integer.}
}
\value{
Returns a list of 'chopped up' data.frames
}
\description{
This function splits the data.frame. The data is very sensitive to its
specific input. Only to be called via harmonizR()
}
