% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotting_missing_values.r
\name{spotting_missing_values}
\alias{spotting_missing_values}
\title{Spotting}
\usage{
spotting_missing_values(main_data, batch_list, needed_values)
}
\arguments{
\item{main_data}{This is the input data.frame read in by the HarmonizR.}

\item{batch_list}{An overview of the batch groupings in list form.}

\item{needed_values}{The number of values needed to be present in a batch
in order to be valid.}
}
\value{
A list of vectors to pass to the upcoming splitting() function.
}
\description{
This function spots missing values within the given data.frame.
}
